refactor: profit, judgement를 participant 내부로 통합
