+++ src/main/java/blackjack/domain/participant/Judgement.java
@@ -1,7 +1,4 @@
-package blackjack.domain.judgement;
-
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
+package blackjack.domain.participant;
 
 public class Judgement {
