+++ src/test/java/blackjack/domain/participant/JudgementTest.java
@@ -1,13 +1,10 @@
-package blackjack.domain.judgement;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardShape;
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Money;
-import blackjack.domain.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
