+++ src/test/java/blackjack/domain/participant/ProfitTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.profit;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -9,9 +9,6 @@ import java.util.Map;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardShape;
-import blackjack.domain.participant.Money;
-import blackjack.domain.participant.Participants;
-import blackjack.domain.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
